requirements.txt
redis
lmdb
gunicorn
flask
requests
python-dotenv
